organize stucture and mongodb data base system.

+++ zhunehra/plugins/skip.py
@@ -1,9 +1,12 @@
-from zhunehra.core.module_injector import *
-from zhunehra.plugins.queue import play_next
+from zhunehra.core import clients
+from telethon import events
+from zhunehra.misc.queue import play_next, queues
+
+zhunehra = clients.zhunehra
 
 @zhunehra.on(events.NewMessage(pattern=r"\/skip"))
 async def skip_handler(event):
-    if event.is_group:
+    if event.is_group or event.is_channel:
         user = await event.get_sender()
         try:
             mention = f"[{user.first_name}](tg://user?id={user.id})"
@@ -12,12 +15,15 @@ async def skip_handler(event):
         chat = await event.get_chat()
         chat_id = int(f"-100{chat.id}" if not str(chat.id).startswith("-100") else chat.id)
         status = await event.reply("**Skiping..**")
-        try:
-            await play_next(chat_id)
-            await event.delete()
-        except Exception:
-            pass
-        await status.edit(f"**Skiped succesfully.\nSkiped by:** {mention}")
+        if chat_id in queues and len(queues[chat_id]) > 0:
+            try:
+                await play_next(chat_id)
+                await event.delete()
+            except Exception:
+                pass
+            await status.edit(f"**Skiped succesfully.\nSkiped by:** {mention}")
+        else:
+            await event.reply("Zhunehra is not streaming.")
     else:
         await event.reply("This command only for groups.")
     
@@ -27,9 +33,12 @@ async def callback_skip(event):
     mention = f"[{user.first_name}](tg://user?id={user.id})"
     chat_id = event.chat_id
     status = await event.reply("**Skiping..**")
-    try:
-        await play_next(chat_id)
-        await event.delete()
-    except Exception:
-        pass
-    await status.edit(f"**Skiped succesfully.\nSkiped by:** {mention}")
+    if chat_id in queues and len(queues[chat_id]) > 0:
+        try:
+            await play_next(chat_id)
+            await event.delete()
+        except Exception:
+            pass
+        await status.edit(f"**Skiped succesfully.\nSkiped by:** {mention}")
+    else:
+        await event.reply("Zhunehra is not streaming.")
